requirements.txt
# Core Framework
fastapi==0.120.0
uvicorn[standard]==0.38.0
pydantic==2.12.3
pydantic_core==2.41.4
pydantic[email]==2.12.3

# Database
sqlalchemy==2.0.36
psycopg2-binary==2.9.10  # PostgreSQL adapter
alembic==1.14.0  # Database migrations
python-dotenv==1.0.1  # Environment variables

# Authentication
python-jose[cryptography]==3.3.0  # JWT tokens
passlib[bcrypt]==1.7.4  # Password hashing
python-multipart==0.0.20  # Form data parsing

# CORS & Middleware
starlette==0.48.0

# Utilities
annotated-types==0.7.0
typing_extensions==4.15.0
typing-inspection==0.4.2

# HTTP & Requests
httpie==3.2.4
requests==2.32.5
requests-toolbelt==1.0.0
certifi==2025.10.5
charset-normalizer==3.4.4
idna==3.11
urllib3==2.5.0

# CLI & Terminal
click==8.3.0
colorama==0.4.6
rich==14.2.0
Pygments==2.19.2
markdown-it-py==4.0.0
mdurl==0.1.2

# Network 
h11==0.16.0
anyio==4.11.0
sniffio==1.3.1
multidict==6.7.0
PySocks==1.7.1

# Other
annotated-doc==0.0.2
defusedxml==0.7.1
annotated-doc==0.0.2
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
defusedxml==0.7.1
fastapi==0.120.0
h11==0.16.0
httpie==3.2.4
idna==3.11
markdown-it-py==4.0.0
mdurl==0.1.2
multidict==6.7.0
pydantic==2.12.3
pydantic_core==2.41.4
Pygments==2.19.2
PySocks==1.7.1
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.2.0
sniffio==1.3.1
starlette==0.48.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.38.0
